backend/app/requirements.txt
alembic
asyncpg
celery[redis]
email-validator
fastapi
httpx
passlib[argon2]
pika
pydantic[email]
pydantic-settings
python-jose[cryptography]
python-multipart
redis
sqlalchemy
uvicorn[standard]
gunicorn
